Added ConfirmPassword. Ran auto formatter on all web files.

+++ web/index.js
@@ -3,10 +3,11 @@ window.onload = function () {
   var signInUserName = document.getElementById("signInUserName");
   var signInPassword = document.getElementById("signInPassword");
   var errorMessage = document.getElementById("errorMessage");
-  
+
   var signUpForm = document.getElementById("signupForm");
   var signUpUserName = document.getElementById("userName");
   var signUpPassword = document.getElementById("psw");
+  var confirmPassword = document.getElementById("confirmPassword");
   var errorMessageSignUp = document.getElementById("errorMessageSignUp");
 
   signInForm.onsubmit = function (e) {
@@ -29,25 +30,31 @@ window.onload = function () {
       )}&password=${encodeURIComponent(signInPassword.value)}`
     );
   };
+
+  signUpForm.onsubmit = function (e) {
+    e.preventDefault();
+    errorMessageSignUp.innerText = "";
     
-  signUpForm.onsubmit = function(e) {
-      e.preventDefault();
-      errorMessageSignUp.innerText = "";
-      var request = new XMLHttpRequest();
-      request.open("POST", "./SignUpServlet", true);
-      request.onload = function () {
-        if (request.status === 200) {
-          sessionStorage.setItem("userName", signUpUserName.value);
-          sessionStorage.setItem("password", signUpPassword.value);
-          window.location = "./home.html";
-        } else {
-          errorMessageSignUp.innerText = request.response;
-        }
-      };
-      request.send(
-        `username=${encodeURIComponent(
-          signUpUserName.value
-        )}&password=${encodeURIComponent(signUpPassword.value)}`
-      );
+    if (confirmPassword.value !== signUpPassword.value) {
+        errorMessageSignUp.innerText = "Passwords don't match.";
+        return;
+    }
+    
+    var request = new XMLHttpRequest();
+    request.open("POST", "./SignUpServlet", true);
+    request.onload = function () {
+      if (request.status === 200) {
+        sessionStorage.setItem("userName", signUpUserName.value);
+        sessionStorage.setItem("password", signUpPassword.value);
+        window.location = "./home.html";
+      } else {
+        errorMessageSignUp.innerText = request.response;
+      }
+    };
+    request.send(
+      `username=${encodeURIComponent(
+        signUpUserName.value
+      )}&password=${encodeURIComponent(signUpPassword.value)}`
+    );
   };
 };
